changes to ui displaying monthly cost
operations changed to return int instead of double for max occupancy

+++ src/B_proj/Operations.java
@@ -153,7 +153,7 @@ public class Operations {
 	 *
 	 * @return the max occupancy
 	 */
-	public double getMaxOccupancy() {
+	public int getMaxOccupancy() {
 		return maxOccupancy;
 	}
